feat(ecosystem): refactoring api into base class
