+++ src/sentry/shared_integrations/constants.py
@@ -1,0 +1,10 @@
+from __future__ import absolute_import
+
+
+ERR_INTERNAL = (
+    "An internal error occurred with the integration and the Sentry team has" " been notified"
+)
+
+ERR_UNAUTHORIZED = "Unauthorized: either your access token was invalid or you do not have" " access"
+
+ERR_UNSUPPORTED_RESPONSE_TYPE = "An unsupported response type was returned: {content_type}"
