+++ src/sentry_plugins/bitbucket/repository_provider.py
@@ -9,7 +9,7 @@ from sentry.models import OrganizationOption
 from sentry.plugins import providers
 from sentry.utils.http import absolute_uri
 
-from sentry_plugins.exceptions import ApiError
+from sentry.shared_integrations.exceptions import ApiError
 
 from .endpoints.webhook import parse_raw_user_email, parse_raw_user_name
 from .mixins import BitbucketMixin
