+++ src/sentry_plugins/client.py
@@ -1,245 +1,18 @@
 from __future__ import absolute_import
 
-from collections import OrderedDict
-import logging
-import json
-import requests
-import six
 
-from bs4 import BeautifulSoup
-from django.utils.functional import cached_property
-from requests.exceptions import ConnectionError, Timeout, HTTPError
-
-from sentry.http import build_session
-from sentry.utils import metrics
-
-from .exceptions import (
-    ApiHostError,
-    ApiTimeoutError,
-    ApiError,
-    ApiUnauthorized,
-    UnsupportedResponseType,
-)
+from sentry.shared_integrations.client import BaseApiClient
+from sentry.shared_integrations.exceptions import ApiUnauthorized
 
 
-class BaseApiResponse(object):
-    text = ""
+class ApiClient(BaseApiClient):
+    integration_type = "plugin"
 
-    def __init__(self, headers=None, status_code=None):
-        self.headers = headers
-        self.status_code = status_code
+    datadog_prefix = "sentry-plugins"
 
-    def __repr__(self):
-        return u"<%s: code=%s, content_type=%s>" % (
-            type(self).__name__,
-            self.status_code,
-            self.headers.get("Content-Type", "") if self.headers else "",
-        )
-
-    @cached_property
-    def rel(self):
-        if not self.headers:
-            return {}
-        link_header = self.headers.get("Link")
-        if not link_header:
-            return {}
-        return {item["rel"]: item["url"] for item in requests.utils.parse_header_links(link_header)}
-
-    @classmethod
-    def from_response(self, response, allow_text=False):
-        # XXX(dcramer): this doesnt handle leading spaces, but they're not common
-        # paths so its ok
-        if response.text.startswith(u"<?xml"):
-            return XmlApiResponse(response.text, response.headers, response.status_code)
-        elif response.text.startswith("<"):
-            if not allow_text:
-                raise ValueError("Not a valid response type: {}".format(response.text[:128]))
-            elif response.status_code < 200 or response.status_code >= 300:
-                raise ValueError(
-                    "Received unexpected plaintext response for code {}".format(
-                        response.status_code
-                    )
-                )
-            return TextApiResponse(response.text, response.headers, response.status_code)
-
-        # Some APIs will return JSON with an invalid content-type, so we try
-        # to decode it anyways
-        if "application/json" not in response.headers["Content-Type"]:
-            try:
-                data = json.loads(response.text, object_pairs_hook=OrderedDict)
-            except (TypeError, ValueError):
-                if allow_text:
-                    return TextApiResponse(response.text, response.headers, response.status_code)
-                raise UnsupportedResponseType(
-                    response.headers["Content-Type"], response.status_code
-                )
-        else:
-            data = json.loads(response.text, object_pairs_hook=OrderedDict)
-
-        if isinstance(data, dict):
-            return MappingApiResponse(data, response.headers, response.status_code)
-        elif isinstance(data, (list, tuple)):
-            return SequenceApiResponse(data, response.headers, response.status_code)
-        else:
-            raise NotImplementedError
-
-
-class TextApiResponse(BaseApiResponse):
-    def __init__(self, text, *args, **kwargs):
-        self.text = text
-        super(TextApiResponse, self).__init__(*args, **kwargs)
-
-
-class XmlApiResponse(BaseApiResponse):
-    def __init__(self, text, *args, **kwargs):
-        self.xml = BeautifulSoup(text, "xml")
-        super(XmlApiResponse, self).__init__(*args, **kwargs)
-
-
-class MappingApiResponse(dict, BaseApiResponse):
-    def __init__(self, data, *args, **kwargs):
-        dict.__init__(self, data)
-        BaseApiResponse.__init__(self, *args, **kwargs)
-
-    @property
-    def json(self):
-        return self
-
-
-class SequenceApiResponse(list, BaseApiResponse):
-    def __init__(self, data, *args, **kwargs):
-        list.__init__(self, data)
-        BaseApiResponse.__init__(self, *args, **kwargs)
-
-    @property
-    def json(self):
-        return self
-
-
-class ApiClient(object):
-    base_url = None
-
-    allow_text = False
-
-    allow_redirects = None
-
-    logger = logging.getLogger("sentry.plugins")
+    log_path = "sentry.plugins"
 
     plugin_name = "undefined"
-
-    def __init__(self, verify_ssl=True, logging_context=None):
-        self.verify_ssl = verify_ssl
-        self.logging_context = logging_context
-
-    def track_response_data(self, code, error=None):
-        logger = logging.getLogger("sentry.plugins.client")
-
-        metrics.incr(
-            "sentry-plugins.http_response",
-            sample_rate=1.0,
-            tags={"plugin": self.plugin_name, "status": code},
-        )
-
-        extra = {
-            "plugin": self.plugin_name,
-            "status_string": six.text_type(code),
-            "error": six.text_type(error)[:128] if error else None,
-        }
-        extra.update(getattr(self, "logging_context", None) or {})
-        logger.info("plugins.http_response", extra=extra)
-
-    def build_url(self, path):
-        if path.startswith("/"):
-            if not self.base_url:
-                raise ValueError("Invalid URL: {}".format(path))
-            return "{}{}".format(self.base_url, path)
-        return path
-
-    def _request(
-        self,
-        method,
-        path,
-        headers=None,
-        data=None,
-        params=None,
-        auth=None,
-        json=True,
-        allow_text=None,
-        allow_redirects=None,
-        timeout=None,
-    ):
-
-        if allow_text is None:
-            allow_text = self.allow_text
-
-        if allow_redirects is None:
-            allow_redirects = self.allow_redirects
-
-        if allow_redirects is None:  # is still None
-            allow_redirects = method.upper() == "GET"
-
-        if timeout is None:
-            timeout = 30
-
-        full_url = self.build_url(path)
-        metrics.incr("sentry-plugins.http_request", tags={"plugin": self.plugin_name})
-
-        session = build_session()
-        try:
-            resp = getattr(session, method.lower())(
-                url=full_url,
-                headers=headers,
-                json=data if json else None,
-                data=data if not json else None,
-                params=params,
-                auth=auth,
-                verify=self.verify_ssl,
-                allow_redirects=allow_redirects,
-                timeout=timeout,
-            )
-            resp.raise_for_status()
-        except ConnectionError as e:
-            self.track_response_data("connection_error", e)
-            raise ApiHostError.from_exception(e)
-        except Timeout as e:
-            self.track_response_data("timeout", e)
-            raise ApiTimeoutError.from_exception(e)
-        except HTTPError as e:
-            resp = e.response
-            if resp is None:
-                self.track_response_data("unknown", e)
-                self.logger.exception(
-                    "request.error", extra={"plugin": self.plugin_name, "url": full_url}
-                )
-                raise ApiError("Internal Error")
-            self.track_response_data(resp.status_code, e)
-            raise ApiError.from_response(resp)
-
-        self.track_response_data(resp.status_code)
-
-        if resp.status_code == 204:
-            return {}
-
-        return BaseApiResponse.from_response(resp, allow_text=allow_text)
-
-    # subclasses should override ``request``
-    def request(self, *args, **kwargs):
-        return self._request(*args, **kwargs)
-
-    def delete(self, *args, **kwargs):
-        return self.request("DELETE", *args, **kwargs)
-
-    def get(self, *args, **kwargs):
-        return self.request("GET", *args, **kwargs)
-
-    def patch(self, *args, **kwargs):
-        return self.request("PATCH", *args, **kwargs)
-
-    def post(self, *args, **kwargs):
-        return self.request("POST", *args, **kwargs)
-
-    def put(self, *args, **kwargs):
-        return self.request("PUT", *args, **kwargs)
 
 
 class AuthApiClient(ApiClient):
