+++ src/sentry_plugins/gitlab/client.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 from six.moves.urllib.parse import quote
 from sentry_plugins.client import ApiClient
 
-from sentry_plugins.exceptions import ApiError
+from sentry.shared_integrations.exceptions import ApiError
 
 
 class GitLabClient(ApiClient):
