+++ tests/sentry_plugins/test_client.py
@@ -6,7 +6,7 @@ import responses
 from sentry.utils.compat.mock import Mock
 from sentry.testutils import TestCase
 
-from sentry_plugins.exceptions import (
+from sentry.shared_integrations.exceptions import (
     ApiError,
     ApiHostError,
     ApiUnauthorized,
